Added MovieDisplayActivity
Made changes to MovieListActivity to pass in the selected Movie index to a a new Activity (MovieDisplayActivity) to display information about the movie. Currently only passes an int, should be changed to an actual Movie object. Note that passing the Movie would require that object to implement Parcelable.
Also added the same code to UserListActivity to display a list of all Users (not actual User objects yet).